requirements.txt
boto3==1.34.47
botocore==1.34.48
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
elasticsearch==7.17.8
emoji==2.10.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jmespath==1.0.1
pyspellchecker==0.8.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
requests-aws4auth==1.2.3
rfc3986==1.5.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
urllib3==1.26.18
